Split shared/result_bundle.py into a layered package
result_bundle.py was 1,201 lines covering canonical naming, hashing,
provenance capture, manifest write/verify, evidence cross-checks, and the
promotion audit. Split it into submodules that layer bottom-up:

  constants.py   schema versions, filenames, recorded dependency set
  errors.py      ResultBundleError
  naming.py      canonical algorithm/backend/path spellings
  hashing.py     content hashes and deterministic JSON writing
  provenance.py  capture and update run provenance
  manifest.py    write and verify the artifact manifest
  evidence.py    summary vs CSV vs per-episode evidence cross-checks
  audit.py       the whole-bundle pre-promotion audit

The package __init__ re-exports the same public surface. The two tests
that patched _repository_state / _dependency_versions now patch the
provenance module that owns them, and the _sanitize_repository_url unit
test imports it from there.

Also run ruff format over the new reporting/ package, which had not been
formatted when it was created.
