+++ environments/shared/result_bundle/errors.py
@@ -1,0 +1,5 @@
+"""Exception raised when a result bundle is invalid or incomplete."""
+
+
+class ResultBundleError(ValueError):
+    """Raised when a result bundle is incomplete or internally inconsistent."""
